python/courier/client/requirements.txt
python-dotenv==1.0.0
requests==2.26.0
sentry-sdk==1.32.0